refactor(vllm_model): fold audio sidechannel into vllm_model, drop vllm_audio_model
Per upstream review feedback: a 30-line metadata-audio-url splice doesn't
need its own model type. Moved the splice (and its 6 unit tests) into
vllm_model._preprocess_chat_completion_create_params. The splice
short-circuits when responses_create_params.metadata.audio_url is absent,
so non-audio benchmarks see no behavior change.

Removed:
- responses_api_models/vllm_audio_model/ (the entire dir)

Updated:
- responses_api_models/vllm_model/app.py: splice logic appended to the
  existing _preprocess_chat_completion_create_params override path.
- responses_api_models/vllm_model/tests/test_app.py: added
  TestAudioSidechannelSplice with the 6 cases from vllm_audio_model.
- benchmarks/librispeech_pc/{config.yaml is unchanged} +
  recipe scripts + Gym-side READMEs all reference vllm_model directly.
