--- responses_api_models/vllm_audio_model/configs/vllm_audio_model.yaml
@@ -1,12 +1,0 @@
-policy_model:
-  responses_api_models:
-    vllm_audio_model:
-      entrypoint: app.py
-      base_url: ${policy_base_url}
-      api_key: ${policy_api_key}
-      model: ${policy_model_name}
-      return_token_id_information: false
-      uses_reasoning_parser: false
-      uses_interleaved_reasoning: false
-      chat_template_kwargs: null
-      extra_body: null
